Adopt SQLite interface/implementation crates similar to key-value
