+++ crates/sqlite-inproc/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-name = "spin-sqlite"
+name = "spin-sqlite-inproc"
 version = { workspace = true }
 authors = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-spin-core = { path = "../core" }
-spin-app = { path = "../app" }
-spin-key-value = { path = "../key-value" }
+spin-sqlite = { path = "../sqlite" }
 spin-world = { path = "../world" }
 anyhow = "1.0"
-wit-bindgen-wasmtime = { workspace = true }
+rusqlite = { version = "0.29.0", features = [ "bundled" ] }
+rand = "0.8"
+once_cell = "1"
 tokio = "1"
