+++ crates/sqlite/src/host_component.rs
@@ -1,58 +1,9 @@
-use std::{
-    collections::HashMap,
-    path::PathBuf,
-    sync::{Arc, Mutex},
-};
+use std::{collections::HashMap, sync::Arc};
 
-use super::Connection;
-use once_cell::sync::OnceCell;
+use crate::{ConnectionManager, SqliteDispatch};
 use spin_app::{AppComponent, DynamicHostComponent};
 use spin_core::HostComponent;
 use spin_world::sqlite;
-
-use crate::SqliteImpl;
-
-#[derive(Debug, Clone)]
-pub enum DatabaseLocation {
-    InMemory,
-    Path(PathBuf),
-}
-
-/// A connection to a sqlite database
-pub struct SqliteConnection {
-    location: DatabaseLocation,
-    connection: OnceCell<Arc<dyn Connection>>,
-}
-
-impl SqliteConnection {
-    pub fn new(location: DatabaseLocation) -> Self {
-        Self {
-            location,
-            connection: OnceCell::new(),
-        }
-    }
-}
-
-impl ConnectionManager for SqliteConnection {
-    fn get_connection(&self) -> Result<Arc<dyn Connection>, sqlite::Error> {
-        let connection = self
-            .connection
-            .get_or_try_init(|| -> Result<_, sqlite::Error> {
-                let c = match &self.location {
-                    DatabaseLocation::InMemory => rusqlite::Connection::open_in_memory(),
-                    DatabaseLocation::Path(path) => rusqlite::Connection::open(path),
-                }
-                .map_err(|e| sqlite::Error::Io(e.to_string()))?;
-                Ok(Arc::new(Mutex::new(c)))
-            })?
-            .clone();
-        Ok(connection)
-    }
-}
-
-pub trait ConnectionManager: Send + Sync {
-    fn get_connection(&self) -> Result<Arc<dyn Connection + 'static>, sqlite::Error>;
-}
 
 pub struct SqliteComponent {
     connection_managers: HashMap<String, Arc<dyn ConnectionManager>>,
@@ -67,7 +18,7 @@ impl SqliteComponent {
 }
 
 impl HostComponent for SqliteComponent {
-    type Data = super::SqliteImpl;
+    type Data = super::SqliteDispatch;
 
     fn add_to_linker<T: Send>(
         linker: &mut spin_core::Linker<T>,
@@ -77,7 +28,7 @@ impl HostComponent for SqliteComponent {
     }
 
     fn build_data(&self) -> Self::Data {
-        SqliteImpl::new(self.connection_managers.clone())
+        SqliteDispatch::new(self.connection_managers.clone())
     }
 }
