+++ crates/trigger/Cargo.toml
@@ -22,7 +22,8 @@ spin-key-value-azure = { path = "../key-value-azure" }
 spin-key-value-redis = { path = "../key-value-redis" }
 spin-key-value-sqlite = { path = "../key-value-sqlite" }
 spin-sqlite = { path = "../sqlite" }
-spin-sqlite-turso = { path = "../sqlite-turso" }
+spin-sqlite-inproc = { path = "../sqlite-inproc" }
+spin-sqlite-libsql = { path = "../sqlite-libsql" }
 sanitize-filename = "0.4"
 serde = "1.0"
 serde_json = "1.0"
